+++ src/index.ts
@@ -3,11 +3,10 @@ import * as brorand from 'brorand'
 import * as hashjs from 'hash.js'
 import * as elliptic from 'elliptic'
 
-import * as addressCodec from 'ripple-address-codec'
+import * as addressCodec from 'brt-address-codec'
 import { derivePrivateKey, accountPublicFromPublicGenerator } from './secp256k1'
 import * as utils from './utils'
 
-const Ed25519 = elliptic.eddsa('ed25519')
 const Secp256k1 = elliptic.ec('secp256k1')
 
 const { hexToBytes } = utils
@@ -16,21 +15,15 @@ const { bytesToHex } = utils
 function generateSeed(
   options: {
     entropy?: Uint8Array
-    algorithm?: 'ed25519' | 'ecdsa-secp256k1'
   } = {},
 ): string {
   assert(!options.entropy || options.entropy.length >= 16, 'entropy too short')
   const entropy = options.entropy ? options.entropy.slice(0, 16) : brorand(16)
-  const type = options.algorithm === 'ed25519' ? 'ed25519' : 'secp256k1'
-  return addressCodec.encodeSeed(entropy, type)
+  return addressCodec.encodeSeed(entropy)
 }
 
 function hash(message): number[] {
-  return hashjs
-    .sha512()
-    .update(message)
-    .digest()
-    .slice(0, 32)
+  return hashjs.sha512().update(message).digest().slice(0, 32)
 }
 
 const secp256k1 = {
@@ -44,10 +37,7 @@ const secp256k1 = {
     const prefix = '00'
 
     const privateKey =
-      prefix +
-      derivePrivateKey(entropy, options)
-        .toString(16, 64)
-        .toUpperCase()
+      prefix + derivePrivateKey(entropy, options).toString(16, 64).toUpperCase()
 
     const publicKey = bytesToHex(
       Secp256k1.keyFromPrivate(privateKey.slice(2))
@@ -70,42 +60,9 @@ const secp256k1 = {
   },
 }
 
-const ed25519 = {
-  deriveKeypair(
-    entropy: Uint8Array,
-  ): {
-    privateKey: string
-    publicKey: string
-  } {
-    const prefix = 'ED'
-    const rawPrivateKey = hash(entropy)
-    const privateKey = prefix + bytesToHex(rawPrivateKey)
-    const publicKey =
-      prefix + bytesToHex(Ed25519.keyFromSecret(rawPrivateKey).pubBytes())
-    return { privateKey, publicKey }
-  },
-
-  sign(message, privateKey): string {
-    // caution: Ed25519.sign interprets all strings as hex, stripping
-    // any non-hex characters without warning
-    assert(Array.isArray(message), 'message must be array of octets')
-    return bytesToHex(
-      Ed25519.sign(message, hexToBytes(privateKey).slice(1)).toBytes(),
-    )
-  },
-
-  verify(message, signature, publicKey): boolean {
-    return Ed25519.verify(
-      message,
-      hexToBytes(signature),
-      hexToBytes(publicKey).slice(1),
-    )
-  },
-}
-
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 function select(algorithm): any {
-  const methods = { 'ecdsa-secp256k1': secp256k1, ed25519 }
+  const methods = { 'ecdsa-secp256k1': secp256k1 }
   return methods[algorithm]
 }
 
@@ -117,7 +74,7 @@ function deriveKeypair(
   privateKey: string
 } {
   const decoded = addressCodec.decodeSeed(seed)
-  const algorithm = decoded.type === 'ed25519' ? 'ed25519' : 'ecdsa-secp256k1'
+  const algorithm = 'ecdsa-secp256k1'
   const method = select(algorithm)
   const keypair = method.deriveKeypair(decoded.bytes, options)
   const messageToVerify = hash('This test message should verify.')
@@ -129,20 +86,13 @@ function deriveKeypair(
   return keypair
 }
 
-function getAlgorithmFromKey(key): 'ed25519' | 'ecdsa-secp256k1' {
-  const bytes = hexToBytes(key)
-  return bytes.length === 33 && bytes[0] === 0xed
-    ? 'ed25519'
-    : 'ecdsa-secp256k1'
-}
-
 function sign(messageHex, privateKey): string {
-  const algorithm = getAlgorithmFromKey(privateKey)
+  const algorithm = 'ecdsa-secp256k1'
   return select(algorithm).sign(hexToBytes(messageHex), privateKey)
 }
 
 function verify(messageHex, signature, publicKey): boolean {
-  const algorithm = getAlgorithmFromKey(publicKey)
+  const algorithm = 'ecdsa-secp256k1'
   return select(algorithm).verify(hexToBytes(messageHex), signature, publicKey)
 }